Populate Coordinates in History Fragment

+++ Alpha/app/src/main/java/com/example/rkuch/alpha/HistoryLogEntry.java
@@ -2,15 +2,15 @@ package com.example.rkuch.alpha;
 
 // TODO: Use double for latitude and longitude
 public class HistoryLogEntry {
-    private int timestamp;
+    private long timestamp;
     private String locationText;
 
-    public HistoryLogEntry(int timestamp, String locationText) {
+    public HistoryLogEntry(long timestamp, String locationText) {
         this.timestamp = timestamp;
         this.locationText = locationText;
     }
 
-    public int getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
